[UPDATE] update package name & remove mongodb
