--- server/chat-service/cmd/main.go
@@ -1,75 +1,0 @@
-package main
-
-import (
-	"os"
-	"os/signal"
-	"reflect"
-	"syscall"
-
-	"github.com/joho/godotenv"
-	log "github.com/sirupsen/logrus"
-	"github.com/widcraft/chat-service/internal/adapter/driven/persistence/client"
-	"github.com/widcraft/chat-service/internal/adapter/driven/persistence/repository"
-	"github.com/widcraft/chat-service/internal/adapter/driving/grpc"
-	"github.com/widcraft/chat-service/internal/adapter/driving/rest"
-	"github.com/widcraft/chat-service/internal/application/service"
-	"github.com/widcraft/chat-service/internal/application/service/message"
-)
-
-var logger = log.New()
-
-type Closable interface {
-	Close() error
-}
-
-// env
-func init() {
-	err := godotenv.Load()
-	if err != nil {
-		panic(err)
-	}
-}
-
-func main() {
-	mongoDb, err := client.NewMongoDb(client.MongoDbConfig{
-		User:     os.Getenv("MONGODB_USER"),
-		Password: os.Getenv("MONGODB_PASSWORD"),
-		Host:     os.Getenv("MONGODB_HOST"),
-		Port:     os.Getenv("MONGODB_PORT"),
-	})
-	if err != nil {
-		logger.Error(err)
-		return
-	}
-
-	chatService := service.NewChatService(
-		logger,
-		message.NewMessageService(logger, repository.NewMessageRepository(mongoDb)),
-		message.NewMessengerService(logger),
-	)
-
-	restApp := rest.New(logger, chatService)
-	grpcApp := grpc.New(logger, chatService)
-
-	defer shutdown(restApp, grpcApp)
-
-	go restApp.Run(os.Getenv("REST_PORT"))
-	go grpcApp.Run(os.Getenv("GRPC_PORT"))
-
-	terminationChan := make(chan os.Signal, 1)
-	signal.Notify(terminationChan, os.Interrupt, syscall.SIGQUIT, syscall.SIGINT, syscall.SIGTERM)
-	<-terminationChan
-}
-
-func shutdown(closables ...Closable) {
-	for _, closable := range closables {
-		if reflect.ValueOf(closable).IsNil() {
-			continue
-		}
-
-		err := closable.Close()
-		if err != nil {
-			logger.Errorf("%s closing failed: %s", reflect.TypeOf(closable), err)
-		}
-	}
-}
