--- server/chat-service/docker-compose.yml
@@ -1,139 +1,0 @@
-version: "3"
-
-services:
-  mongodb:
-    image: mongo
-    container_name: chat_mongodb
-    ports:
-      - ${MONGODB_PORT}:${MONGODB_PORT}
-    volumes:
-      - mongodb-data:/data/db
-    environment:
-      MONGO_INITDB_ROOT_USERNAME: ${MONGODB_USER}
-      MONGO_INITDB_ROOT_PASSWORD: ${MONGODB_PASSWORD}
-      MONGO_INITDB_DATABASE: ${MONGODB_DB}
-      MONGO_NON_ROOT_USERNAME: ${MONGODB_USER}
-      MONGO_NON_ROOT_PASSWORD: ${MONGODB_PASSWORD}
-      MONGO_NON_ROOT_DATABASE: ${MONGODB_AUTH_DB}
-    networks:
-      - mongo_network
-
-  mongo-express:
-    image: mongo-express
-    ports:
-      - 8081:8081
-    environment:
-      ME_CONFIG_MONGODB_ADMINUSERNAME: ${MONGODB_USER}
-      ME_CONFIG_MONGODB_ADMINPASSWORD: ${MONGODB_PASSWORD}
-      ME_CONFIG_MONGODB_URL: mongodb://${MONGODB_USER}:${MONGODB_PASSWORD}@mongodb:${MONGODB_PORT}
-      ME_CONFIG_BASICAUTH_USERNAME: ${MONGODB_USER}
-      ME_CONFIG_BASICAUTH_PASSWORD: ${MONGODB_PASSWORD}
-    networks:
-      - mongo_network
-
-  Kafka00Service:
-    image: bitnami/kafka:3.5.1-debian-11-r44
-    restart: unless-stopped
-    container_name: Kafka00Container
-    ports:
-      - "10000:9094"
-    environment:
-      - KAFKA_CFG_BROKER_ID=0
-      - KAFKA_CFG_NODE_ID=0
-      - KAFKA_KRAFT_CLUSTER_ID=HsDBs9l6UUmQq7Y5E6bNlw
-      - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=0@Kafka00Service:9093,1@Kafka01Service:9093,2@Kafka02Service:9093
-      - ALLOW_PLAINTEXT_LISTENER=yes
-      - KAFKA_CFG_AUTO_CREATE_TOPICS_ENABLE=true
-      - KAFKA_CFG_LISTENERS=PLAINTEXT://:9092,CONTROLLER://:9093,EXTERNAL://:9094
-      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT://Kafka00Service:9092,EXTERNAL://127.0.0.1:10000
-      - KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP=CONTROLLER:PLAINTEXT,EXTERNAL:PLAINTEXT,PLAINTEXT:PLAINTEXT
-      - KAFKA_CFG_OFFSETS_TOPIC_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_MIN_ISR=2
-      - KAFKA_CFG_PROCESS_ROLES=controller,broker
-      - KAFKA_CFG_CONTROLLER_LISTENER_NAMES=CONTROLLER
-    networks:
-      - kafka_network
-    volumes:
-      - Kafka00:/bitnami/kafka
-
-  Kafka01Service:
-    image: bitnami/kafka:3.5.1-debian-11-r44
-    restart: unless-stopped
-    container_name: Kafka01Container
-    ports:
-      - "10001:9094"
-    environment:
-      - KAFKA_CFG_BROKER_ID=1
-      - KAFKA_CFG_NODE_ID=1
-      - KAFKA_KRAFT_CLUSTER_ID=HsDBs9l6UUmQq7Y5E6bNlw
-      - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=0@Kafka00Service:9093,1@Kafka01Service:9093,2@Kafka02Service:9093
-      - ALLOW_PLAINTEXT_LISTENER=yes
-      - KAFKA_CFG_AUTO_CREATE_TOPICS_ENABLE=true
-      - KAFKA_CFG_LISTENERS=PLAINTEXT://:9092,CONTROLLER://:9093,EXTERNAL://:9094
-      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT://Kafka01Service:9092,EXTERNAL://127.0.0.1:10001
-      - KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP=CONTROLLER:PLAINTEXT,EXTERNAL:PLAINTEXT,PLAINTEXT:PLAINTEXT
-      - KAFKA_CFG_OFFSETS_TOPIC_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_MIN_ISR=2
-      - KAFKA_CFG_PROCESS_ROLES=controller,broker
-      - KAFKA_CFG_CONTROLLER_LISTENER_NAMES=CONTROLLER
-    networks:
-      - kafka_network
-    volumes:
-      - Kafka01:/bitnami/kafka
-
-  Kafka02Service:
-    image: bitnami/kafka:3.5.1-debian-11-r44
-    restart: unless-stopped
-    container_name: Kafka02Container
-    ports:
-      - "10002:9094"
-    environment:
-      - KAFKA_CFG_BROKER_ID=2
-      - KAFKA_CFG_NODE_ID=2
-      - KAFKA_KRAFT_CLUSTER_ID=HsDBs9l6UUmQq7Y5E6bNlw
-      - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=0@Kafka00Service:9093,1@Kafka01Service:9093,2@Kafka02Service:9093
-      - ALLOW_PLAINTEXT_LISTENER=yes
-      - KAFKA_CFG_AUTO_CREATE_TOPICS_ENABLE=true
-      - KAFKA_CFG_LISTENERS=PLAINTEXT://:9092,CONTROLLER://:9093,EXTERNAL://:9094
-      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT://Kafka02Service:9092,EXTERNAL://127.0.0.1:10002
-      - KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP=CONTROLLER:PLAINTEXT,EXTERNAL:PLAINTEXT,PLAINTEXT:PLAINTEXT
-      - KAFKA_CFG_OFFSETS_TOPIC_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_REPLICATION_FACTOR=3
-      - KAFKA_CFG_TRANSACTION_STATE_LOG_MIN_ISR=2
-      - KAFKA_CFG_PROCESS_ROLES=controller,broker
-      - KAFKA_CFG_CONTROLLER_LISTENER_NAMES=CONTROLLER
-    networks:
-      - kafka_network
-    volumes:
-      - Kafka02:/bitnami/kafka
-
-  KafkaWebUiService:
-    image: provectuslabs/kafka-ui:latest
-    restart: always
-    container_name: KafkaWebUiContainer
-    ports:
-      - "8085:8080"
-    environment:
-      - KAFKA_CLUSTERS_0_NAME=Local-Kraft-Cluster
-      - KAFKA_CLUSTERS_0_BOOTSTRAPSERVERS=Kafka00Service:9092,Kafka01Service:9092,Kafka02Service:9092
-      - DYNAMIC_CONFIG_ENABLED=true
-      - KAFKA_CLUSTERS_0_AUDIT_TOPICAUDITENABLED=true
-      - KAFKA_CLUSTERS_0_AUDIT_CONSOLEAUDITENABLED=true
-    depends_on:
-      - Kafka00Service
-      - Kafka01Service
-      - Kafka02Service
-    networks:
-      - kafka_network
-
-networks:
-  mongo_network:
-  kafka_network:
-
-volumes:
-  mongodb-data:
-  Kafka00:
-  Kafka01:
-  Kafka02:
