--- server/chat-service/internal/adapter/driven/persistence/client/mongo_db.go
@@ -1,30 +1,0 @@
-package client
-
-import (
-	"context"
-	"fmt"
-
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-)
-
-type MongoDbConfig struct {
-	User     string
-	Password string
-	Host     string
-	Port     string
-}
-
-type MongoDb struct {
-	*mongo.Client
-}
-
-func NewMongoDb(config MongoDbConfig) (*MongoDb, error) {
-	applyUri := fmt.Sprintf("mongodb://%s:%s@%s:%s", config.User, config.User, config.Host, config.Port)
-	client, err := mongo.Connect(context.TODO(), options.Client().ApplyURI(applyUri))
-	if err != nil {
-		return nil, err
-	}
-
-	return &MongoDb{client}, nil
-}
