--- server/chat-service/internal/adapter/driving/rest/handler/chatroom/handler.go
@@ -1,7 +1,0 @@
-package chatroom
-
-import "github.com/widcraft/chat-service/internal/port/driven"
-
-type Handler struct {
-	logger driven.Logger
-}
