--- server/chat-service/internal/application/abstraction/message_service.go
@@ -1,8 +1,0 @@
-package abstraction
-
-import "github.com/widcraft/chat-service/internal/application/dto"
-
-type MessageService interface {
-	SaveMessage(message *dto.MessageDto) error          // NOTE: should move to kafka message consumer pipeline
-	GetMessages(roomIdx uint) ([]dto.MessageDto, error) // TODO: cursor pagination functionality
-}
