--- server/chat-service/internal/port/driven/message_port.go
@@ -1,8 +1,0 @@
-package driven
-
-import "github.com/widcraft/chat-service/internal/domain/entity"
-
-type MessageRepository interface {
-	SaveMessage(message *entity.Message) error
-	GetMessages(roomIdx uint) ([]entity.Message, error)
-}
