--- server/chat-service/internal/port/driving/message_port.go
@@ -1,10 +1,0 @@
-package driving
-
-import "github.com/widcraft/chat-service/internal/application/dto"
-
-type ChatService interface {
-	Join(client MessengerClient)
-	Leave(client MessengerClient)
-	SendMessage(messageDto *dto.MessageDto) error
-	GetMessages(roomIdx uint) ([]dto.MessageDto, error)
-}
