--- server/chat-service/internal/port/driving/messenger_port.go
@@ -1,9 +1,0 @@
-package driving
-
-import "github.com/widcraft/chat-service/internal/application/dto"
-
-type MessengerClient interface {
-	GetRoomIdx() uint
-	GetUserIdx() uint
-	SendMessage(message *dto.MessageDto) error
-}
