+++ server/messenger-service/internal/adapter/driving/rest/handler/chat/payload.go
