+++ server/messenger-service/internal/adapter/driving/rest/handler/chatroom/handler.go
@@ -1,0 +1,7 @@
+package chatroom
+
+import "github.com/widcraft/messenger-service/internal/port/driven"
+
+type Handler struct {
+	logger driven.Logger
+}
