+++ server/messenger-service/internal/application/dto/message_dto.go
