+++ server/messenger-service/internal/domain/entity/message.go
