+++ server/messenger-service/internal/domain/entity/room.go
