+++ server/messenger-service/internal/port/driven/message_producer_port.go
@@ -1,0 +1,7 @@
+package driven
+
+import "github.com/widcraft/messenger-service/internal/application/dto"
+
+type MessageProducerPort interface {
+	ProduceMessage(message *dto.MessageDto) error
+}
